*Test.cpp, TestHelper.hpp: Changed C++ test runners to count using size_t instead of unsigned int, tweaked test helper code to support this change and simplify internal code.

+++ cpp/Base58CheckTest.cpp
@@ -182,7 +182,7 @@ static void testPublicAddressExport() {
 		{false, "A3571B72C6914434AB0F6C4EBDA1472FEBA88DEF", "1FtfVasxiAgW1JgsbndQgs48EKJiXHHVCGa"},
 	};
 	
-	for (unsigned int i = 0; i < ARRAY_LENGTH(cases); i++) {
+	for (size_t i = 0; i < ARRAY_LENGTH(cases); i++) {
 		TestCase &tc = cases[i];
 		Bytes pubkeyHash(hexBytes(tc.hexadecimal));
 		assert(pubkeyHash.size() == 20);
@@ -349,7 +349,7 @@ static void testPublicAddressImport() {
 		{true, "FEEDBFF7DFF6BDF7FFFD6FFFEF7FFBEFFFFFFF7F", "1QEwSsoa4tYREcCm7YVYa9RH6V8rkqsV4F"},
 	};
 	
-	for (unsigned int i = 0; i < ARRAY_LENGTH(cases); i++) {
+	for (size_t i = 0; i < ARRAY_LENGTH(cases); i++) {
 		TestCase &tc = cases[i];
 		Bytes expected(hexBytes(tc.hexadecimal));
 		assert(expected.size() == 20);
@@ -516,7 +516,7 @@ static void testPrivateKeyExport() {
 		{false, "9DCBD84EB62DF0BB2094184F64258B86494511787E5AEE41EDA9A95D39ED4358", "L2WSpW5cxTUP3dbjHiuh4gDvxF4WFu7dGfuTh4VMZauqY7C41o5"},
 	};
 	
-	for (unsigned int i = 0; i < ARRAY_LENGTH(cases); i++) {
+	for (size_t i = 0; i < ARRAY_LENGTH(cases); i++) {
 		TestCase &tc = cases[i];
 		assert(strlen(tc.hexadecimal) == 64);
 		char actual[53];
@@ -685,7 +685,7 @@ static void testPrivateKeyImport() {
 		{true, "FD7FF6FFFF7FE7FBFDFFDFBFF77FDFFDFFF3BFFFFFE7FFD7FFFD7BBEFFEEFB7F", "L5iUsypV2eizXzvFFaNuMvkvQsmZ3nZTjDXyYAX6CHKfD2FFQmgX"},
 	};
 	
-	for (unsigned int i = 0; i < ARRAY_LENGTH(cases); i++) {
+	for (size_t i = 0; i < ARRAY_LENGTH(cases); i++) {
 		TestCase &tc = cases[i];
 		assert(strlen(tc.hexadecimal) == 64);
 		Uint256 privKey;
